# Fast table-driven parsing for upb (2+GB/s)