--- a/BUILD
+++ b/BUILD
@@ -79,6 +79,7 @@ cc_library(
     name = "upb",
     srcs = [
         "upb/decode.c",
+        "upb/decode_fast.c",
         "upb/encode.c",
         "upb/msg.c",
         "upb/msg.h",