--- a/BUILD
+++ b/BUILD
@@ -26,12 +26,6 @@ exports_files([
 ])
 
 config_setting(
-    name = "darwin",
-    values = {"cpu": "darwin"},
-    visibility = ["//visibility:public"],
-)
-
-config_setting(
     name = "windows",
     constraint_values = ["@bazel_tools//platforms:windows"],
 )