--- a/benchmark.py
+++ b/benchmark.py
@@ -30,12 +30,12 @@ def Run(cmd):
 def Benchmark(outbase, runs=12):
   tmpfile = "/tmp/bench-output.json"
   Run("rm -rf {}".format(tmpfile))
-  Run("bazel test :all")
-  Run("bazel build -c opt :benchmark")
+  Run("CC=clang bazel test :all")
+  Run("CC=clang bazel build -c opt :benchmark")
 
   Run("./bazel-bin/benchmark --benchmark_out_format=json --benchmark_out={} --benchmark_repetitions={}".format(tmpfile, runs))
 
-  Run("bazel build -c opt --copt=-g :conformance_upb")
+  Run("CC=clang bazel build -c opt --copt=-g :conformance_upb")
   Run("cp -f bazel-bin/conformance_upb {}.bin".format(outbase))
 
   with open(tmpfile) as f: