--- a/cmake/build_defs.bzl
+++ b/cmake/build_defs.bzl
@@ -1,4 +1,3 @@
-
 def generated_file_staleness_test(name, outs, generated_pattern):
     """Tests that checked-in file(s) match the contents of generated file(s).