--- a/cmake/make_cmakelists.py
+++ b/cmake/make_cmakelists.py
@@ -141,6 +141,9 @@ class BuildFileFunctions(object):
   def config_setting(self, **kwargs):
     pass
 
+  def upb_fasttable_enabled(self, **kwargs):
+    pass
+
   def select(self, arg_dict):
     return []