--- a/cmake/make_cmakelists.py
+++ b/cmake/make_cmakelists.py
@@ -129,6 +129,9 @@ class BuildFileFunctions(object):
   def upb_proto_library(self, **kwargs):
     pass
 
+  def upb_proto_library_copts(self, **kwargs):
+    pass
+
   def upb_proto_reflection_library(self, **kwargs):
     pass