--- a/kokoro/ubuntu/build.sh
+++ b/kokoro/ubuntu/build.sh
@@ -17,7 +17,6 @@ which bazel
 bazel version
 
 cd $(dirname $0)/../..
-<<<<<<< HEAD
 
 if which gcc; then
   gcc --version
@@ -31,13 +30,12 @@ if which clang; then
   # The checked-in code is with fasttable not enabled.
   CC=clang bazel test --test_output=errors ... --//:fasttable_enabled=true -- -cmake:test_generated_files
 fi
-=======
->>>>>>> master
 
 if which gcc; then
   gcc --version
   CC=gcc bazel test --test_output=errors ...
   CC=gcc bazel test -c opt --test_output=errors ...
+  CC=gcc bazel test --test_output=errors ... --//:fasttable_enabled=true -- -cmake:test_generated_files
   # TODO: work through these errors and enable this.
   # if gcc -fanalyzer -x c /dev/null -c -o /dev/null; then
   #   CC=gcc bazel test --copt=-fanalyzer --test_output=errors ...
@@ -47,6 +45,7 @@ fi
 if which clang; then
   CC=clang bazel test --test_output=errors ...
   CC=clang bazel test --test_output=errors -c opt ...
+  CC=clang bazel test --test_output=errors ... --//:fasttable_enabled=true -- -cmake:test_generated_files
 
   if [[ $(uname) = "Linux" ]]; then
     CC=clang bazel test --test_output=errors --config=m32 ...