… CMakeLists.txt
… benchmarks/BUILD
… benchmarks/BUILD.googleapis
… benchmarks/compare.py
… cmake/CMakeLists.txt
… cmake/google/protobuf/descriptor.upb.c
… generated_for_cmake/google/protobuf/descriptor.upb.c
… tests/benchmark.cc
… tests/bindings/lua/test_upb.lua
… tests/test_generated_code.c
… third_party/wyhash/BUILD
… upb/decode.int.h
… upb/decode_fast.c
… upb/decode_fast.h
… upb/def.c
… upb/json_decode.c
… upb/msg.h
… upb/port_def.inc
… upbc/generator.cc
… upbc/main.cc
… upbc/message_layout.cc