@davidbolvansky 2021-04-25
Related to your article and "jne followed by jmp" issue - LLVM does it intentionally this way:
https://llvm.org/doxygen/BranchFolding_8cpp_source.html

line 1521